# perf: add fast path to `count_bits_set_by_offsets`