--- a/crates/polars-ops/src/chunked_array/list/count.rs
+++ b/crates/polars-ops/src/chunked_array/list/count.rs
@@ -6,9 +6,20 @@ use arrow::legacy::utils::CustomIterTools;
 use super::*;
 
 fn count_bits_set_by_offsets(values: &Bitmap, offset: &[i64]) -> Vec<IdxSize> {
-    // Fast path where all bits are unset.
+    // Fast path where all bits are either set or unset.
     if values.unset_bits() == values.len() {
         return vec![0 as IdxSize; offset.len() - 1];
+    } else if values.unset_bits() == 0 {
+        let mut start = offset[0];
+        let v = (offset[1..])
+            .iter()
+            .map(|end| {
+                let current_offset = start;
+                start = *end;
+                (end - current_offset) as IdxSize
+            })
+            .collect_trusted();
+        return v;
     }
 
     let (bits, bitmap_offset, _) = values.as_slice();
@@ -23,12 +34,8 @@ fn count_bits_set_by_offsets(values: &Bitmap, offset: &[i64]) -> Vec<IdxSize> {
 
             let len = (end - current_offset) as usize;
 
-            // Fast path where all bits are set.
-            if values.unset_bits() == 0 {
-                return len as IdxSize;
-            }
-
-            (len - count_zeros(bits, bitmap_offset + current_offset as usize, len)) as IdxSize
+            let set_ones = len - count_zeros(bits, bitmap_offset + current_offset as usize, len);
+            set_ones as IdxSize
         })
         .collect_trusted()
 }