--- a/py-polars/tests/unit/datatypes/test_list.py
+++ b/py-polars/tests/unit/datatypes/test_list.py
@@ -349,6 +349,15 @@ def test_list_sum_and_dtypes() -> None:
         "a": [1, 6, 10, 15, None]
     }
 
+    # Booleans
+    assert pl.DataFrame(
+        {"a": [[True], [True, True], [True, False, True], [True, True, True, None]]},
+    ).select(pl.col("a").list.sum()).to_dict(as_series=False) == {"a": [1, 2, 2, 3]}
+
+    assert pl.DataFrame(
+        {"a": [[False], [False, False], [False, False, False]]},
+    ).select(pl.col("a").list.sum()).to_dict(as_series=False) == {"a": [0, 0, 0]}
+
 
 def test_list_mean() -> None:
     assert pl.DataFrame({"a": [[1], [1, 2, 3], [1, 2, 3, 4], [1, 2, 3, 4, 5]]}).select(