Add fast path to `count_bits_set_by_offsets` when all bits are either set or cleared.

Some benchmarks:

```python
# Mixed True's and False's
df = pl.DataFrame({'A': [[False, True] * 100, [True] * 200] * 500})
%timeit df.with_columns(sum=pl.col('A').list.sum())
304 µs ± 2.34 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)


# All False
df = pl.DataFrame({'A': [[False, False] * 100, [False] * 200] * 500})
%timeit df.with_columns(sum=pl.col('A').list.sum())
68.7 µs ± 301 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)


# All True
df = pl.DataFrame({'A': [[True, True] * 100, [True] * 200] * 500})
%timeit df.with_columns(pl.col('A').list.sum())
72.4 µs ± 392 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```